apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coto-api"
version = "0.1.0"
description = "AI English conversation practice API"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "structlog>=24.4",
    "redis>=5.2",
    "openai>=1.60",
    "python-multipart>=0.0.18",
    "sse-starlette>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
    "mypy>=1.14",
    "httpx>=0.28",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]

[tool.ruff.lint.per-file-ignores]
# B008: Depends() in default args is standard FastAPI pattern
"src/coto/routers/*.py" = ["B008"]
# TCH003: SQLAlchemy resolves Mapped[] annotations at runtime via get_type_hints()
"src/coto/models/*.py" = ["TCH003"]

[tool.ruff.lint.isort]
known-first-party = ["coto"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/coto"]
